load times ux improvements

+++ lib/edit_contact.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:brasil_fields/brasil_fields.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
@@ -31,6 +29,7 @@ class _EditContactPageState extends State<EditContactPage> {
   late Contact _contact;
   late ContactRepository _contactRepository;
   bool _isNewContact = false;
+  bool _isSaving = false;
 
   @override
   void initState() {
@@ -53,22 +52,43 @@ class _EditContactPageState extends State<EditContactPage> {
         appBar: AppBar(
           backgroundColor: Theme.of(context).colorScheme.inversePrimary,
           title: Text(widget.title),
-          actions: [
-            GestureDetector(
-                onTap: () async {
-                  bool success = await _contactRepository.delete(_contact);
-                  if (success) {
-                    nameInputController.text = '';
-                    surnameInputController.text = '';
-                    phoneInputController.text = '';
-                    _contact = Contact.empty();
-                    setState(() {});
-                  }
-                },
-                child: Container(
-                    margin: const EdgeInsets.only(right: 20),
-                    child: const Icon(Icons.delete))),
-          ],
+          actions: _isNewContact
+              ? []
+              : [
+                  GestureDetector(
+                      onTap: () async {
+                        if (!_isSaving) {
+                          setState(() {
+                            _isSaving = true;
+                          });
+                          bool success =
+                              await _contactRepository.delete(_contact);
+                          if (success) {
+                            nameInputController.text = '';
+                            surnameInputController.text = '';
+                            phoneInputController.text = '';
+                            _contact = Contact.empty();
+
+                            await _showDialog('Contato apagado com sucesso.');
+
+                            await Future.delayed(const Duration(seconds: 1));
+                            if (!context.mounted) return;
+                            setState(() {
+                              _isSaving = false;
+                            });
+                            Navigator.of(context).pop();
+                          } else {
+                            _showDialog('Erro. Contato não foi apagado.');
+                            setState(() {
+                              _isSaving = false;
+                            });
+                          }
+                        }
+                      },
+                      child: Container(
+                          margin: const EdgeInsets.only(right: 20),
+                          child: const Icon(Icons.delete))),
+                ],
         ),
         body: Container(
             margin: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
@@ -105,6 +125,7 @@ class _EditContactPageState extends State<EditContactPage> {
                         height: 20,
                       ),
                       TextFormField(
+                        keyboardType: TextInputType.name,
                         controller: nameInputController,
                         decoration: const InputDecoration(
                           labelText: 'Nome',
@@ -112,6 +133,7 @@ class _EditContactPageState extends State<EditContactPage> {
                         ),
                       ),
                       TextFormField(
+                        keyboardType: TextInputType.name,
                         controller: surnameInputController,
                         decoration: const InputDecoration(
                           labelText: 'Sobrenome',
@@ -121,7 +143,7 @@ class _EditContactPageState extends State<EditContactPage> {
                       TextFormField(
                         controller: phoneInputController,
                         keyboardType: TextInputType.number,
-                        inputFormatters: [TelefoneInputFormatter()],
+                        //inputFormatters: [TelefoneInputFormatter()],
                         decoration: const InputDecoration(
                           labelText: 'Telefone',
                           icon: Icon(Icons.phone),
@@ -130,37 +152,96 @@ class _EditContactPageState extends State<EditContactPage> {
                       const SizedBox(
                         height: 20,
                       ),
-                      TextButton(
-                          onPressed: () async {
-                            debugPrint('will save...');
-                            _contact.setName(nameInputController.text);
-                            _contact.setSurname(surnameInputController.text);
-                            _contact.setPhone(phoneInputController.text);
-                            //if insert else update
-                            bool success = _isNewContact
-                                ? await _contactRepository.insert(_contact)
-                                : await _contactRepository.update(_contact);
-                            if (success) {
-                              setState(() {});
-                            }
-                          },
-                          style: ButtonStyle(
-                              shape: MaterialStateProperty.all(
-                                  RoundedRectangleBorder(
-                                      borderRadius: BorderRadius.circular(10))),
-                              backgroundColor:
-                                  MaterialStateProperty.all(Colors.red)),
-                          child: const SizedBox(
-                              width: 100,
-                              child: Row(
-                                  mainAxisAlignment: MainAxisAlignment.center,
-                                  children: <Widget>[
-                                    Text(
-                                      'Salvar',
-                                      style: TextStyle(
-                                          color: Colors.white, fontSize: 20),
-                                    )
-                                  ])))
+                      _isSaving
+                          ? CircularProgressIndicator()
+                          : TextButton(
+                              onPressed: () async {
+                                debugPrint('will save...');
+                                _contact.setName(nameInputController.text);
+                                _contact
+                                    .setSurname(surnameInputController.text);
+                                _contact.setPhone(phoneInputController.text);
+                                _contact.setPhoto('photo');
+
+                                setState(() {
+                                  _isSaving = true;
+                                });
+
+                                if (_contact.getName.isNotEmpty &&
+                                    _contact.getSurname.isNotEmpty &&
+                                    _contact.getPhone.isNotEmpty) {
+                                  bool success = _isNewContact
+                                      ? await _contactRepository
+                                          .insert(_contact)
+                                      : await _contactRepository
+                                          .update(_contact);
+                                  if (success) {
+                                    await _showDialog(
+                                        'Contato criado com sucesso.');
+                                    await Future.delayed(
+                                        const Duration(seconds: 1));
+                                    if (!context.mounted) return;
+                                    setState(() {
+                                      _isSaving = false;
+                                    });
+                                    Navigator.of(context).pop();
+                                  } else {
+                                    _showDialog(
+                                        'Erro ao criar contato. Não foi possível adicionar ao banco de dados.');
+                                    setState(() {
+                                      _isSaving = false;
+                                    });
+                                  }
+                                } else {
+                                  _showDialog(
+                                      'Contato não criado. Verifique os campos.');
+                                  setState(() {
+                                    _isSaving = false;
+                                  });
+                                }
+                              },
+                              style: ButtonStyle(
+                                  shape: MaterialStateProperty.all(
+                                      RoundedRectangleBorder(
+                                          borderRadius:
+                                              BorderRadius.circular(10))),
+                                  backgroundColor:
+                                      MaterialStateProperty.all(Colors.red)),
+                              child: const SizedBox(
+                                  width: 100,
+                                  child: Row(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.center,
+                                      children: <Widget>[
+                                        Text(
+                                          'Salvar',
+                                          style: TextStyle(
+                                              color: Colors.white,
+                                              fontSize: 20),
+                                        )
+                                      ])))
                     ])))));
+  }
+
+  Future _showDialog(String message, {String title = 'Alerta'}) {
+    return showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        // retorna um objeto do tipo Dialog
+        return AlertDialog(
+          title: Text(title),
+          content: Text(message),
+          actions: <Widget>[
+            // define os botões na base do dialogo
+            TextButton(
+              child: const Text("Fechar"),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+          ],
+        );
+      },
+    );
   }
 }
